WIP
Cleanup code and fix bugs

+++ src/headers/handle-input.hpp
@@ -16,7 +16,7 @@ using namespace std;
 extern unsigned char **all_images;
 extern unsigned char **query_images;
 
-void handleInput(int , char**, uint32_t*, uint64_t*, int*, int*, int*, double*, string*, string*);
-void handleReExecution(uint32_t*, uint64_t*, int*, int*, int*, double*, string*, string*);
+void handleInput(int , char**, uint32_t*, uint64_t*, int*, int*, int*, unsigned int*, string*, string*);
+void handleReExecution(uint32_t*, uint64_t*, int*, int*, int*, unsigned int*, string*, string*);
 void readFile(const string&, int, uint32_t*, uint64_t*);
 void printFiles(uint32_t number_of_images, uint32_t number_of_query_images, uint64_t d);
