+++ src/headers/process.hpp
@@ -1,1 +1,1 @@
-void process(uint32_t, uint64_t, int, int, int, double);
+void process(uint32_t, uint64_t, int, int, int, unsigned int);
